Cast em "return" corrigindo erro de compilação para a compilação usando o maven. Apenas com o eclipse esse erro não era mostrado (ou era mostrado como warning).
Exclusão de arquivos ".odt" do build do Bopepo.

+++ src/main/java/org/jrimum/bopepo/FatorDeVencimento.java
@@ -141,7 +141,7 @@ public class FatorDeVencimento{
 
 		if (isNull(data)) {
 			
-			return Exceptions.throwIllegalArgumentException("Impossível realizar o cálculo do fator de vencimento de uma data nula!");
+			return (Integer) Exceptions.throwIllegalArgumentException("Impossível realizar o cálculo do fator de vencimento de uma data nula!");
 			
 		} else {
